refactor: separate send and receive test
