+++ CMakeLists.txt
@@ -66,14 +66,17 @@ if(BUILD_TESTING)
   find_package(ament_lint_auto REQUIRED)
 
   ament_add_gtest(${PROJECT_NAME}_tests
+    "test/send_and_receive/raw_test.cpp"
+    "test/send_and_receive/string_test.cpp"
     "test/base_broadcaster_and_listener_test.cpp"
     "test/base_client_and_server_test.cpp"
     "test/broadcaster_and_listener_test.cpp"
     "test/client_and_server_test.cpp"
-    "test/send_and_receive_test.cpp"
     "test/string_broadcaster_and_listener_test.cpp"
     "test/tcp_socket_test.cpp"
     "test/udp_socket_test.cpp")
+
+  target_include_directories(${PROJECT_NAME}_tests PUBLIC "test/include")
 
   target_link_libraries(${PROJECT_NAME}_tests ${PROJECT_NAME})
