+++ test/include/musen_test/mock_sender.hpp
@@ -1,0 +1,57 @@
+// Copyright (c) 2021 ICHIRO ITS
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+// THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+#ifndef MUSEN_TEST__MOCK_SENDER_HPP_
+#define MUSEN_TEST__MOCK_SENDER_HPP_
+
+#include <musen/musen.hpp>
+
+#include <memory>
+#include <queue>
+#include <vector>
+
+namespace musen_test
+{
+
+using Buffers = std::queue<std::vector<char>>;
+
+class MockSender : public musen::Sender
+{
+public:
+  explicit MockSender(std::shared_ptr<Buffers> buffers)
+  : buffers(buffers)
+  {
+  }
+
+  size_t send_raw(const char * data, const size_t & length) override
+  {
+    std::vector buffer(data, data + length);
+    buffers->push(buffer);
+
+    return buffer.size();
+  }
+
+private:
+  std::shared_ptr<Buffers> buffers;
+};
+
+}  // namespace musen_test
+
+#endif  // MUSEN_TEST__MOCK_SENDER_HPP_
